overlay_windows: move to separate file
